Feat: Implementar bordes de imagen en modo híbrido
Añade la funcionalidad de renderizado de bordes de imagen (curvados y biselados) en el modo de generación de PDF híbrido.

Anteriormente, los bordes definidos en las opciones de "Diseño Avanzado" solo se aplicaban en el modo de generación estándar. Esta actualización implementa la lógica necesaria utilizando trazados de recorte (clipping paths) en PDFBox para replicar el mismo efecto visual en el modo híbrido.

Cambios clave:
- Se ha añadido una nueva función de ayuda, 'applyPdfBoxClippingPath', para crear y aplicar trazados de recorte en un 'PDPageContentStream'.
- Las funciones 'drawImagesWithPdfBox' para la portada y las páginas interiores ahora utilizan esta función para aplicar los bordes antes de dibujar la imagen.
- Se ha corregido un error en la carga de imágenes de páginas interiores en modo híbrido, asegurando que se manejen correctamente los 'content URI' de Android.
- Se ha corregido un error de compilación por un tipo de dato incorrecto.

Con esto, la consistencia visual entre el modo estándar y el híbrido está completa.

+++ app/src/main/java/com/example/dynamiccollage/utils/HybridPdfRenderer.kt
@@ -582,7 +582,7 @@ private fun drawImagesWithPdfBox(
             PageOrientation.Vertical -> if (pageData.imageUris.size > 1) Pair(1, 2) else Pair(1, 1)
             PageOrientation.Horizontal -> if (pageData.imageUris.size > 1) Pair(2, 1) else Pair(1, 1)
         }
-        val rects = getRectsForPage(pageWidth, pageHeight.toInt(), startY, cols, rows, 15f)
+        val rects = getRectsForPage(pageWidth.toInt(), pageHeight.toInt(), startY, cols, rows, 15f)
         val borderSettings = coverConfig.imageBorderSettingsMap[pageData.groupId]
 
         pageData.imageUris.forEachIndexed { index, uriString ->
